Support rest parameters
allow the ability to catch all the arguments, not just positional.
This will allow having a CLI that works like so:

  fnm exec --using=12 node -v
                           ^^ - positional argument, can be forwarded

instead of:

  fnm exec --using=12 -- node -v

The problem is that with the current implementation of the parser,
options and circuitbreakers look "everywhere" within the node array
so the following:

  fnm exec --using=12 node --help

will be the same as:

  fnm exec --help --using=12 node

Eventually this needs to be taken care of. A flag can be global, but
the default behavior should be local to the declaration. CLI commands
form a tree and we need to make sure we don't have global searches
by default.